update: Changed GITApi to git-api.ts
