+++ Frontend/src/app/git-api.spec.ts
@@ -1,0 +1,7 @@
+import { GitAPI } from './git-api';
+
+describe('GitAPI', () => {
+  it('should create an instance', () => {
+    expect(new GitAPI()).toBeTruthy();
+  });
+});
